requirements.txt
numpy<2.0
Pillow==9.5.0
PySide6==6.8.3
dataclasses==0.6
safetensors==0.3.1
openai==1.76.0
google-genai
torch==1.13.1
opencv-python==4.11.0.86
-e git+https://github.com/PramaLLC/BEN2.git#egg=ben2
